pyproject.toml
[project]
name = "long-term-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.2.2",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=3.0.3",
    "numpy>=2.4.0",
    "psycopg-pool>=3.3.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
